refactor(test): restructure package structure and update references in user service tests
- Moved test classes to appropriate sub-packages (`crud` and `util`) for better organization.
- Updated exceptions and method names in `UsernameGeneratorServiceTest` and `UserOperationsServiceTest`.
